pyproject.toml
[project]
name = "backend-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "annotated-types>=0.7.0",
    "asgiref>=3.9.1",
    "azure-ai-evaluation>=1.11.0",
    "azure-ai-inference>=1.0.0b9",
    "azure-ai-projects>=1.0.0",
    "azure-ai-textanalytics>=5.3.0",
    "azure-cognitiveservices-speech>=1.45.0",
    "azure-monitor-opentelemetry>=1.7.0",
    "azure-search-documents>=11.5.3",
    "charset-normalizer>=3.4.3",
    "click>=8.2.1",
    "distro>=1.9.0",
    "eventlet>=0.40.3",
    "fastapi>=0.116.1",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "flask-socketio>=5.5.1",
    "gunicorn>=23.0.0",
    "jiter>=0.10.0",
    "pandas>=2.3.2",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.1",
    "python-socketio[client]>=5.13.0",
    "requests>=2.32.5",
    "streamlit>=1.49.1",
    "watchdog>=6.0.0",
]
